tasks/web/simple/deploy/app/requirements.txt
Flask
pyjwt
cryptography